services/pinochet-api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Diego Quintana <daquintanav@gmail.com>"]
readme = "README.md"
packages = [
    { include = "pinochet", from = "src" },
]

[tool.poetry.scripts]
pinochet-api-cli = "pinochet.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
loguru = "^0.7.2"
python-jose = "^3.3.0"
typer = "^0.9.0"
pydantic-settings = "^2.0.3"
psycopg2-binary = "^2.9.9"
tenacity = "^8.2.3"
strawberry-graphql = {extras = ["fastapi"], version = "^0.213.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.5"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
